+++ internal/names/names.go
@@ -3,6 +3,7 @@ package names
 import (
 	"strings"
 
+	"github.com/glasskube/glasskube/api/v1alpha1"
 	"github.com/glasskube/glasskube/internal/controller/ctrlpkg"
 )
 
@@ -13,4 +14,8 @@ func PackageInfoName(pkg ctrlpkg.Package) string {
 		parts = append([]string{spec.PackageInfo.RepositoryName}, parts...)
 	}
 	return escapeResourceName(strings.Join(parts, "--"))
+}
+
+func HelmResourceName(pkg ctrlpkg.Package, manifest *v1alpha1.PackageManifest) string {
+	return strings.Join([]string{pkg.GetName(), manifest.Name}, "-")
 }