+++ pkg/open/open.go
@@ -7,14 +7,18 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"strings"
 
 	"github.com/glasskube/glasskube/api/v1alpha1"
 	"github.com/glasskube/glasskube/internal/clicontext"
 	"github.com/glasskube/glasskube/internal/cliutils"
 	"github.com/glasskube/glasskube/internal/controller/ctrlpkg"
+	"github.com/glasskube/glasskube/internal/names"
 	"github.com/glasskube/glasskube/pkg/future"
 	"github.com/glasskube/glasskube/pkg/manifest"
+	"go.uber.org/multierr"
 	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -87,7 +91,7 @@ func (o *opener) Open(
 			stopCh := make(chan struct{})
 			o.readyCh = append(o.readyCh, readyCh)
 			o.stopCh = append(o.stopCh, stopCh)
-			entrypointFuture, err := o.open(ctx, namespace, e, readyCh, stopCh)
+			entrypointFuture, err := o.open(ctx, pkg, manifest, namespace, e, readyCh, stopCh)
 			if err != nil {
 				o.stop()
 				epName := e.Name
@@ -139,6 +143,8 @@ func (o *opener) stop() {
 
 func (o *opener) open(
 	ctx context.Context,
+	pkg ctrlpkg.Package,
+	manifest *v1alpha1.PackageManifest,
 	namespace string,
 	entrypoint v1alpha1.PackageEntrypoint,
 	readyChannel chan struct{},
@@ -148,7 +154,7 @@ func (o *opener) open(
 		return nil, err
 	}
 
-	svc, err := o.service(ctx, namespace, entrypoint)
+	svc, err := o.service(ctx, pkg, manifest, namespace, entrypoint)
 	if err != nil {
 		return nil, err
 	}
@@ -187,17 +193,41 @@ func (o *opener) open(
 
 func (o *opener) service(
 	ctx context.Context,
+	pkg ctrlpkg.Package,
+	manifest *v1alpha1.PackageManifest,
 	namespace string,
 	entrypoint v1alpha1.PackageEntrypoint,
 ) (*corev1.Service, error) {
-	svc, err := o.ksClient.CoreV1().
-		Services(namespace).
-		Get(ctx, entrypoint.ServiceName, metav1.GetOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("could not get service %v: %w", entrypoint.ServiceName, err)
-	} else {
-		return svc, nil
+	svcNameCandidates := []string{
+		entrypoint.ServiceName,
 	}
+
+	if manifest.Helm != nil {
+		svcNameCandidates = append(svcNameCandidates,
+			strings.Join([]string{pkg.GetName(), entrypoint.Name}, "-"),
+			strings.Join([]string{names.HelmResourceName(pkg, manifest), entrypoint.Name}, "-"),
+			names.HelmResourceName(pkg, manifest),
+		)
+	}
+
+	var errs error
+
+	for _, name := range svcNameCandidates {
+		svc, err := o.ksClient.CoreV1().
+			Services(namespace).
+			Get(ctx, name, metav1.GetOptions{})
+		if err != nil {
+			if !apierrors.IsNotFound(err) {
+				return nil, fmt.Errorf("could not get service %v: %w", entrypoint.ServiceName, err)
+			} else {
+				multierr.AppendInto(&errs, err)
+			}
+		} else {
+			return svc, nil
+		}
+	}
+
+	return nil, fmt.Errorf("could not find service: %w", errs)
 }
 
 func (o *opener) pod(ctx context.Context, service *corev1.Service) (*corev1.Pod, error) {